Investment/T0_Optimized/requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
akshare==1.10.2
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0
